requirements.txt
# PE parsing
pefile>=2023.2.7

# ELF parsing (detector)
pyelftools>=0.30

# step0_find_and_download_samples.py (Malware Bazaar)
requests>=2.28.0
pyzipper>=0.3.5

# Optional: ASPack unpacking (emulation via Unipacker; needs setuptools for pkg_resources on Python 3.12+)
# unipacker
# setuptools>=60,<70  # if ModuleNotFoundError: pkg_resources

# Optional: emulation for generic unpacker (install separately if needed)
# qiling>=1.4.5
# speakeasy-emulator>=2.0.0

# Config (optional)
# PyYAML>=6.0
